сделал сборку html через gulp с минификацией
