+++ README.md
@@ -1,2 +1,16 @@
 # IR Remote Control Emmeci
 ![Alt text](https://github.com/Igor-kor/IrRemoteEmmeci/blob/master/img/screenshot1.png?raw=true "Screenshot")
+
+install node js (https://nodejs.org/en/)
+install gulp (https://gulpjs.com/docs/en/getting-started/quick-start/)
+npm install --global gulp-cli
+npm install --save-dev gulp
+install minify (https://www.npmjs.com/package/gulp-htmlmin)
+npm install --save gulp-htmlmin
+install twig (https://www.npmjs.com/package/gulp-twig)
+npm install gulp-twig --save
+install minifyjs (https://www.npmjs.com/package/gulp-minify)
+npm install --save-dev gulp-minify
+
+for run 'gulp'
+
