+++ html/source/style.css
@@ -1,0 +1,200 @@
+html {
+    font-family: 'Open Sans', sans-serif;
+    display: block;
+    margin: 0 auto;
+    text-align: center;
+    color: #CCCCCC;
+}
+body {
+    background-color: #333333;
+    margin-top: 50px;
+}
+
+.side-by-side {
+    display: inline-block;
+    vertical-align: middle;
+    position: relative;
+}
+
+.humidity-icon {
+    background-color: #3498db;
+    width: 30px;
+    height: 30px;
+    border-radius: 50%;
+    line-height: 36px;
+}
+
+.humidity-text {
+    font-weight: 600;
+    padding-left: 15px;
+    font-size: 19px;
+    width: 160px;
+    text-align: left;
+}
+
+.humidity {
+    font-weight: 500;
+    font-size: 60px;
+    color: #3498db;
+}
+
+.temperature-icon {
+    background-color: #f39c12;
+    width: 30px;
+    height: 30px;
+    border-radius: 50%;
+    line-height: 40px;
+}
+
+.temperature-text {
+    font-weight: 600;
+    padding-left: 15px;
+    font-size: 19px;
+    width: 160px;
+    text-align: left;
+}
+
+.temperature {
+    font-weight: 500;
+    font-size: 60px;
+    color: #f39c12;
+}
+
+.superscript {
+    font-size: 17px;
+    font-weight: 600;
+    position: absolute;
+    right: -20px;
+    top: 15px;
+}
+
+.data {
+    padding: 10px;
+}
+
+.btn {
+    display: block;
+    vertical-align: middle;
+    position: relative;
+    width: 150px;
+    margin: 10px auto;
+    padding: 15px 34px;
+    overflow: hidden;
+    font-family: 'Montserrat Alternates', sans-serif;
+    color: white;
+    letter-spacing: 1px;
+    text-transform: uppercase;
+    font-weight: 600;
+    border-width: 0;
+    transform: matrix3d(1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1);
+    outline: none;
+    cursor: pointer;
+}
+
+.btn1 {
+    background: linear-gradient(90deg, #c600b2, #7619fd);
+}
+
+.btn2 {
+    background: linear-gradient(90deg, #4764b3, #00e6ff);
+}
+
+.btn3 {
+    background: linear-gradient(90deg, #feff36, #cc481f);
+}
+
+.btn4 {
+    background: linear-gradient(90deg, #c77474, #b224d4);
+}
+
+.btn5 {
+    background: linear-gradient(90deg, #ff1a1a, #d8ba46);
+}
+.btn-save {
+    background: linear-gradient(90deg, #454545, #212121);
+}
+
+.btn:before, .btn:after {
+    content: "";
+    position: absolute;
+    z-index: -1;
+    width: 100%;
+    height: 100%;
+    left: -110%;
+    top: 0;
+    opacity: 0.1;
+    background: black;
+    transform: skewX(25deg);
+    transition: .5s ease;
+}
+
+.btn:hover:before {
+    left: -10%;
+    transition: .3s ease;
+}
+
+.btn:hover:after {
+    left: -20%;
+}
+
+.setting {
+    display: inline-block;
+    position: absolute;
+    right: 30px;
+    margin-top: -20px;
+}
+
+.modal {
+    display: none;
+    position: fixed;
+    z-index: 1;
+    left: 0;
+    top: 0;
+    width: 100%;
+    height: 100%;
+    overflow: auto;
+    background-color: rgb(0, 0, 0);
+    background-color: rgba(0, 0, 0, 0.9);
+}
+
+.modal-content {
+    background-color: #292929;
+    margin: auto auto;
+    width: 90%;
+    min-height: 90%;
+    text-align: left;
+}
+
+.close {
+    color: #aaa;
+    float: right;
+    font-size: 28px;
+    font-weight: bold;
+}
+
+.close:hover,
+.close:focus {
+    color: black;
+    text-decoration: none;
+    cursor: pointer;
+}
+
+.modal-center {
+    text-align: center;
+}
+
+.modal-form > label {
+    display: block;
+    margin-top: 20px;
+}
+
+.modal-form > input {
+    display: block;
+}
+
+.btn-ok > button {
+    bottom: 0;
+}
+#Layer_4{
+    display: none;
+}
